requirements.txt
sentence-transformers==2.2.2
langchain
gunicorn
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain-google-genai
langchain_experimental
-e .
